conf: Seperated graph and contributor cards into components, seperated metrics into a separate file and moved dashboard to a seperate folder
